+++ vrp-core/src/construction/enablers/schedule_update.rs
@@ -10,9 +10,9 @@ pub struct ScheduleKeys {
     pub latest_arrival: StateKey,
     /// Waiting time state key.
     pub waiting_time: StateKey,
-    /// Total distance state key.
+    /// Total route distance state key.
     pub total_distance: StateKey,
-    /// Total duration state key.
+    /// Total route duration state key.
     pub total_duration: StateKey,
     /// Limit duration state key.
     pub limit_duration: StateKey,
@@ -101,10 +101,14 @@ fn update_states(
         0_f64,
     );
 
-    let (route, state) = route_ctx.as_mut();
+    let route = route_ctx.route();
+    let mut latest_arrivals = Vec::with_capacity(route.tour.total());
+    let mut waiting_times = Vec::with_capacity(route.tour.total());
 
-    route.tour.all_activities().enumerate().rev().fold(init, |acc, (activity_idx, act)| {
+    route.tour.all_activities().rev().fold(init, |acc, act| {
         if act.job.is_none() {
+            latest_arrivals.push(Default::default());
+            waiting_times.push(Default::default());
             return acc;
         }
 
@@ -118,11 +122,23 @@ fn update_states(
         };
         let future_waiting = waiting + (act.place.time.start - act.schedule.arrival).max(0.);
 
-        state.put_activity_state(state_keys.latest_arrival, activity_idx, latest_arrival_time);
-        state.put_activity_state(state_keys.waiting_time, activity_idx, future_waiting);
+        latest_arrivals.push(latest_arrival_time);
+        waiting_times.push(future_waiting);
 
         (latest_arrival_time, act.place.location, future_waiting)
     });
+
+    latest_arrivals.reverse();
+    waiting_times.reverse();
+
+    // NOTE: pop out state for arrival
+    if route.tour.end().map_or(false, |end| end.job.is_none()) {
+        latest_arrivals.pop();
+        waiting_times.pop();
+    }
+
+    route_ctx.state_mut().put_activity_states(state_keys.latest_arrival, latest_arrivals);
+    route_ctx.state_mut().put_activity_states(state_keys.waiting_time, waiting_times);
 }
 
 fn update_statistics(
@@ -137,16 +153,11 @@ fn update_statistics(
     let total_dur = end.schedule.departure - start.schedule.departure;
 
     let init = (start.place.location, start.schedule.departure, Distance::default());
-    let (_, _, total_dist) =
-        route.tour.all_activities().enumerate().skip(1).fold(init, |(loc, dep, total_dist), (a_idx, a)| {
-            let total_dist = total_dist + transport.distance(route, loc, a.place.location, TravelTime::Departure(dep));
-            let total_dur = a.schedule.departure - start.schedule.departure;
+    let (_, _, total_dist) = route.tour.all_activities().skip(1).fold(init, |(loc, dep, total_dist), a| {
+        let total_dist = total_dist + transport.distance(route, loc, a.place.location, TravelTime::Departure(dep));
 
-            state.put_activity_state(state_keys.total_distance, a_idx, total_dist);
-            state.put_activity_state(state_keys.total_duration, a_idx, total_dur);
-
-            (a.place.location, a.schedule.departure, total_dist)
-        });
+        (a.place.location, a.schedule.departure, total_dist)
+    });
 
     state.put_route_state(state_keys.total_distance, total_dist);
     state.put_route_state(state_keys.total_duration, total_dur);