+++ vrp-core/tests/helpers/models/solution/route.rs
@@ -108,16 +108,6 @@ pub struct RouteStateBuilder {
 }
 
 impl RouteStateBuilder {
-    pub fn add_activity_state<T: Send + Sync + 'static>(
-        &mut self,
-        key: StateKey,
-        activity_idx: usize,
-        value: T,
-    ) -> &mut Self {
-        self.state.put_activity_state(key, activity_idx, value);
-        self
-    }
-
     pub fn add_route_state<T: Send + Sync + 'static>(&mut self, key: StateKey, value: T) -> &mut Self {
         self.state.put_route_state(key, value);
         self